Extract SelectionList header
